feat: remaining bss changes

+++ app/objects/beatmap.py
@@ -23,6 +23,7 @@ from app.constants.gamemodes import GameMode
 from app.logging import Ansi
 from app.logging import log
 from app.repositories import maps as maps_repo
+from app.repositories.maps import MapServer
 from app.utils import escape_enum
 from app.utils import pymysql_encode
 
@@ -303,6 +304,7 @@ class Beatmap:
         hp: float = 0.0,
         diff: float = 0.0,
         filename: str = "",
+        server: MapServer = MapServer.OSU,
     ) -> None:
         self.set = map_set
 
@@ -328,6 +330,7 @@ class Beatmap:
         self.hp = hp
         self.diff = diff
         self.filename = filename
+        self.server = server
 
     def __repr__(self) -> str:
         return self.full_name
@@ -763,25 +766,29 @@ class BeatmapSet:
             # - should we delete the beatmap from the database if it's not in the osu!api?
             # - are 404 and 200 the only cases where we should delete the beatmap?
             if self.maps:
-                map_md5s_to_delete = {bmap.md5 for bmap in self.maps}
+                map_md5s_to_delete = {
+                    bmap.md5 for bmap in self.maps if bmap.server == MapServer.OSU
+                }
 
-                # delete maps
+                if map_md5s_to_delete:
+                    # delete maps
+                    await app.state.services.database.execute(
+                        "DELETE FROM maps WHERE md5 IN :map_md5s",
+                        {"map_md5s": map_md5s_to_delete},
+                    )
+
+                    # delete scores on the maps
+                    await app.state.services.database.execute(
+                        "DELETE FROM scores WHERE map_md5 IN :map_md5s",
+                        {"map_md5s": map_md5s_to_delete},
+                    )
+
+            if self.maps[0].server == MapServer.OSU:
+                # delete set
                 await app.state.services.database.execute(
-                    "DELETE FROM maps WHERE md5 IN :map_md5s",
-                    {"map_md5s": map_md5s_to_delete},
+                    "DELETE FROM mapsets WHERE id = :set_id",
+                    {"set_id": self.id},
                 )
-
-                # delete scores on the maps
-                await app.state.services.database.execute(
-                    "DELETE FROM scores WHERE map_md5 IN :map_md5s",
-                    {"map_md5s": map_md5s_to_delete},
-                )
-
-            # delete set
-            await app.state.services.database.execute(
-                "DELETE FROM mapsets WHERE id = :set_id",
-                {"set_id": self.id},
-            )
 
     async def _save_to_sql(self) -> None:
         """Save the object's attributes into the database."""
@@ -875,6 +882,7 @@ class BeatmapSet:
                 diff=row["diff"],
                 filename=row["filename"],
                 map_set=bmap_set,
+                server=MapServer(row["server"]),
             )
 
             # XXX: tempfix for bancho.py <v3.4.1,
